Externalize Schema for GET requests
